% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Birthday_Function.r
\name{birthday}
\alias{birthday}
\title{Birthday Function}
\usage{
birthday(x)
}
\arguments{
\item{x}{Integer or Vector, Number of people in group.}
}
\value{
Values between 0 and 1 giving the Probability
}
\description{
Birthday Function
}
\examples{
birthday(20)
birthday(20:24)

}
