% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.r
\docType{data}
\name{ddt}
\alias{ddt}
\title{DDT Date in fish in the Tennessee River and Tributaries}
\format{
\subsection{\code{who}}{

A data frame with 144 rows and 6 columns:
\describe{
\item{country}{Country name}
\item{iso2, iso3}{2 & 3 letter ISO country codes}
\item{year}{Year}
...
}
}
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\usage{
ddt
}
\description{
A subset of data from the World Health Organization Global Tuberculosis
Report ...
}
\keyword{datasets}
