% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.r
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Damage Data}
\format{
\subsection{\code{fire}}{

A data frame with 15 rows and 2 columns:
\describe{
\item{Date}{Date of the fire event}
\item{Location}{Location of the fire (city or region)}
}
}
}
\source{
FIREDAM.csv or your data source
}
\usage{
fire
}
\description{
A dataset of fire damage events, containing details on location and date.
}
\keyword{datasets}
