% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labfunctions.r, R/labfuntctions.r
\name{mybin}
\alias{mybin}
\title{Simulate a Binomial Distribution}
\usage{
mybin(iter = 100, n = 10, p = 0.5)

mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{Number of simulations to run}

\item{n}{Number of trials per simulation}

\item{p}{Probability of success for each trial}
}
\value{
A vector of relative frequencies for 0:n successes

A vector of relative frequencies for 0:n successes
}
\description{
This function simulates a binomial experiment for a given number of iterations,
sample size, and probability of success.

This function simulates a binomial experiment for a given number of iterations,
sample size, and probability of success.
}
\examples{
mybin(iter = 100, n = 10, p = 0.7)
mybin(iter = 100, n = 10, p = 0.7)
}
