% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labfunctions.r, R/labfuntctions.r
\name{myhyper}
\alias{myhyper}
\title{Simulate a Hypergeometric Distribution}
\usage{
myhyper(iter = 100, n, N, r)

myhyper(iter = 100, n, N, r)
}
\arguments{
\item{iter}{Number of simulations to run}

\item{n}{Sample size}

\item{N}{Population size}

\item{r}{Number of successes in the population}
}
\value{
A vector of relative frequencies for 0 to n successes

A vector of relative frequencies for 0 to n successes
}
\description{
This function simulates sampling from a finite population without replacement
and calculates the relative frequencies of the number of successes in each sample.

This function simulates sampling from a finite population without replacement
and calculates the relative frequencies of the number of successes in each sample.
}
\examples{
myhyper(iter = 1000, n = 5, N = 20, r = 12)
myhyper(iter = 1000, n = 5, N = 20, r = 12)
}
