% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.r
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{Mean of the normal distribution}

\item{sigma}{Standard deviation of the normal distribution}

\item{a}{Upper limit for shading}
}
\value{
A list with mu, sigma, and probability P(X <= a)
}
\description{
myncurve
}
\examples{
myncurve(mu = 10, sigma = 5, a = 6)
}
